requirements.txt
# Dependencies for SIA (Safety via Intent Awareness) Implementation

# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.37.0
qwen-vl-utils>=0.0.5

# Image processing
Pillow>=10.0.0

# Progress tracking
tqdm>=4.66.0

# Data handling
numpy>=1.24.0

# Optional but recommended for faster inference
accelerate>=0.25.0
